refactor: Change custom error file
