+++ src/services/api/index.tsx
@@ -1,2 +1,2 @@
 export { axiosProductInstance, axiosUserInstance } from './api';
-export { CustomError } from './api.type';
+export { CustomError } from './CustomError';
